bot/requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
opencv-python==4.9.0.80
mediapipe==0.10.9
Pillow==10.2.0
python-dotenv==1.0.0
redis==5.0.1
aiofiles==23.2.1
numpy==1.26.3
requests==2.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
